Sponsors: Rename file to match paradigm in other blocks
