+++ public_html/wp-content/mu-plugins/blocks/assets/src/sponsors/sponsors-select.js
